Refactor: separate components related `Task`
